+++ CHANGELOG.md
@@ -13,6 +13,10 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 ## v3.0.0-rc.11 (2020-05-10)
 
+#### :rocket: New Feature
+
+* [Added `forceUpdate` parameter for `@field` decorator](https://github.com/V4Fire/Client/pull/226)
+
 #### :bug: Bug Fix
 
 * Fixed calling `window.scrollTo` without window context