+++ CHANGELOG.md
@@ -11,6 +11,16 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.0.0-rc.12 (2020-05-13)
+
+#### :rocket: New Feature
+
+* Added `--browsers` flag `gulp test:component`
+
+#### :bug: Bug Fix
+
+* Fixed initializing of functional components
+
 ## v3.0.0-rc.11 (2020-05-10)
 
 #### :rocket: New Feature